move EnsimeVFS into its own package
