+++ core/src/it/scala/org/ensime/core/RichPresentationCompilerSpec.scala
@@ -7,7 +7,7 @@ import java.io.File
 import akka.event.slf4j.SLF4JLogging
 import org.ensime.api._
 import org.ensime.fixture._
-import org.ensime.indexer.EnsimeVFS
+import org.ensime.vfs._
 import org.scalatest._
 import org.ensime.util.file._
