+++ core/src/it/scala/org/ensime/model/SourcePositionSpec.scala
@@ -5,7 +5,7 @@ package org.ensime.model
 import org.ensime.api._
 import org.ensime.fixture._
 import org.ensime.indexer.DatabaseService.FqnSymbol
-import org.ensime.indexer.EnsimeVFS
+import org.ensime.vfs._
 import org.scalatest._
 import org.ensime.util.file._
