+++ core/src/main/scala/org/ensime/core/Analyzer.scala
@@ -8,7 +8,8 @@ import java.nio.charset.Charset
 import akka.actor._
 import akka.event.LoggingReceive.withLabel
 import org.ensime.api._
-import org.ensime.indexer.{ EnsimeVFS, SearchService }
+import org.ensime.vfs._
+import org.ensime.indexer.SearchService
 import org.ensime.model._
 import org.ensime.util.{ PresentationReporter, ReportHandler, FileUtils }
 import org.slf4j.LoggerFactory
