+++ core/src/main/scala/org/ensime/indexer/FileWatchers.scala
@@ -7,13 +7,13 @@ import org.apache.commons.vfs2._
 import org.apache.commons.vfs2.impl.DefaultFileMonitor
 
 import org.ensime.api._
+import org.ensime.vfs._
 
 import scala.collection.Set
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
 import scala.util.Try
 import org.ensime.util.file._
-import EnsimeVFS._
 
 trait FileChangeListener {
   def fileAdded(f: FileObject): Unit
@@ -68,7 +68,7 @@ class SourceWatcher(
       module <- config.modules.values
       root <- module.sourceRoots
     } yield {
-      new ApachePollingFileWatcher(root, EnsimeVFS.SourceSelector, true, listeners)
+      new ApachePollingFileWatcher(root, SourceSelector, true, listeners)
     }
   override def shutdown(): Unit = impls.foreach(_.shutdown)
 }
